refactor(test): move provision keycloak command to test/integration
The keycloak provisioner is test infrastructure, not a platform
requirement. Move it out of the service binary into a standalone tool
alongside provision-fixtures. Update CI workflows, composite action,
and documentation to use the new invocation pattern.

- service/cmd/provisionKeycloak.go → test/integration/cmd/provision-keycloak/
- Delete service/cmd/provision.go (no longer needed)
- CI uses: GOWORK=off go run -C test/integration ./cmd/provision-keycloak
- Update docs: Contributing.md, Consuming.md, service/README.md
