+++ .github/workflows/nightly-checks.yaml
@@ -44,9 +44,9 @@ jobs:
         working-directory: platform
       - run: docker compose up -d --wait --wait-timeout 240
         working-directory: platform
-      - run: go run ./service provision keycloak
+      - run: GOWORK=off go run -C test/integration ./cmd/provision-keycloak
         working-directory: platform
-      - run: go run ./service provision fixtures
+      - run: GOWORK=off go run -C test/integration ./cmd/provision-fixtures
         working-directory: platform
       - uses: JarvusInnovations/background-action@2428e7b970a846423095c79d43f759abf979a635 # v1.0.7
         name: start server in background
